requirements.txt
amqp==2.4.1
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-filter==2.1.0
django-haystack==2.8.1
django-redis==4.10.0
djangorestframework==3.9.2
drf-extensions==0.4.0
elasticsearch==6.3.1
flower==0.9.2
idna==2.8
kombu==4.3.0
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
qiniu==7.2.2
qrcode==6.1
redis==3.2.0
reportlab==3.5.13
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
tornado==5.1.1
urllib3==1.24.1
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
